Split bip_conf_enum into different files
